refactor: replace AppException with proper HTTP exception types

+++ src/Controller/SgfsController.php
@@ -37,7 +37,7 @@ class SgfsController extends AppController
 		if (isset($this->params['url']['delete']))
 		{
 			$sDel = $this->Sgf->findById($this->params['url']['delete']);
-			if (Auth::getUserID() == $sDel['Sgf']['user_id'])
+			if ($sDel && Auth::getUserID() == $sDel['Sgf']['user_id'])
 				$this->Sgf->delete($sDel['Sgf']['id']);
 		}
 
@@ -47,9 +47,15 @@ class SgfsController extends AppController
 			foreach ($newDuplicates as $duplicateId)
 			{
 				$dupl = $this->Tsumego->findById($duplicateId);
+				if (!$dupl)
+					continue;
 				$scT = $this->SetConnection->find('first', ['conditions' => ['tsumego_id' => $dupl['Tsumego']['id']]]);
+				if (!$scT)
+					continue;
 				$dupl['Tsumego']['set_id'] = $scT['SetConnection']['set_id'];
 				$dSet = $this->Set->findById($dupl['Tsumego']['set_id']);
+				if (!$dSet)
+					continue;
 				$dId[] = $dupl['Tsumego']['id'];
 				$dTitle[] = $dSet['Set']['title'] . ' - ' . $dupl['Tsumego']['num'];
 			}
